requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
packaging==24.2
python-decouple==3.8
sqlparse==0.5.3
setuptools
whitenoise==6.9.0